+++ myCryptoWallet-ConsoleApp/src/main/java/com/isa/control/transactions/Transactions.java
@@ -1,0 +1,16 @@
+package com.isa.control.transactions;
+
+import com.isa.control.Coin;
+
+public class Transactions {
+    private Integer idTransaction;
+    private final Coin coin;
+    private boolean isActive;
+    private double volume;
+
+    public Transactions(Coin coin, boolean isActive, double volume) {
+        this.coin = coin;
+        this.isActive = isActive;
+        this.volume = volume;
+    }
+}
